+++ src/taurenmd/libs/libutil.py
@@ -1,4 +1,5 @@
 """General utilities."""
+import os
 import re
 
 
@@ -37,3 +38,46 @@ def split_time_unit(s):
     except IndexError as err:
         raise ValueError(f'Time string not appropriate: {s!r}') from err
     return float(value), unit
+
+
+def make_csv_lines_in_interleaved_manner(data, labels):
+    """
+    Create CSV lines from a list of lists in interleaved manner.
+
+    Labels and data do not need to be of the same size. Resulting
+    sizes will be completed with empty strings for those series with
+    less values.
+
+    Parameters
+    ----------
+    data : list of lists of numbers
+        The data series.
+
+    labels : list of lists of str
+        The labels refering to the data series.
+
+    Returns
+    -------
+    str
+        A string merging the lines.
+    """
+    # require
+    assert all(len(i) == len(j) for i, j in zip(data, labels)), \
+        'data and labels size do not match'
+
+    lines = []
+    max_data = max(len(_r) for _r in data)
+    for i in range(max_data):
+        line = ''
+        for j in range(len(data)):
+
+            try:
+                _l = labels[j][i]
+                _r = data[j][i]
+            except IndexError:
+                _l, _r = '', ''
+
+            line += f'{_l},{_r},'
+
+        lines.append(line.rstrip(','))
+    return os.linesep.join(lines)