+++ tests/test_libutil.py
@@ -1,4 +1,5 @@
 """Test libutil."""
+import os
 from math import isclose
 
 import pytest
@@ -60,3 +61,82 @@ def test_string_to_timestep_wrong(s):
     """Test correct conversion."""
     with pytest.raises(ValueError):
         libutil.split_time_unit(s)
+
+
+def test_interleaved_str_for_csv():
+    """Test making interleaved CSV strings."""
+    indata = [
+        [1, 2, 3, 4],
+        [5, 6, 7, 8],
+        ]
+
+    inlabels = [
+        ['a', 'b', 'c', 'd'],
+        ['e', 'f', 'g', 'h'],
+        ]
+
+    result = libutil.make_csv_lines_in_interleaved_manner(indata, inlabels)
+
+    expected = (
+        f'a,1,e,5{os.linesep}'
+        f'b,2,f,6{os.linesep}'
+        f'c,3,g,7{os.linesep}'
+        'd,4,h,8'
+        )
+
+    assert result == expected
+
+
+def test_interleaved_str_for_csv_2():
+    """Test making interleaved CSV strings."""
+    indata = [
+        [1, 2, 3, 4],
+        [5, 6, 7],
+        ]
+
+    inlabels = [
+        ['a', 'b', 'c', 'd'],
+        ['e', 'f', 'g'],
+        ]
+
+    result = libutil.make_csv_lines_in_interleaved_manner(indata, inlabels)
+
+    expected = (
+        f'a,1,e,5{os.linesep}'
+        f'b,2,f,6{os.linesep}'
+        f'c,3,g,7{os.linesep}'
+        'd,4'
+        )
+
+    assert result == expected
+
+
+@pytest.mark.parametrize(
+    'indata,inlabels',
+    [
+        (
+            [
+                [1, 2, 3, 4],
+                [5, 6, 7],
+                ],
+            [
+                ['a', 'b', 'c', 'd'],
+                ['e', 'f', 'g', 'e'],
+                ]
+            ),
+        (
+            [
+                [1, 2, 3],
+                [5, 6, 7, 8],
+                ],
+            [
+                ['a', 'b', 'c', 'd'],
+                ['e', 'f', 'g'],
+                ]
+            ),
+        ]
+    )
+def test_interleaved_str_for_csv_3(indata, inlabels):
+    """Test making interleaved CSV strings."""
+    with pytest.raises(AssertionError):
+        libutil.make_csv_lines_in_interleaved_manner(indata, inlabels)